debian changes
git-svn-id: svn://oss.syntext.com/oss/serna-free/trunk@268 76c9efe7-43af-4aca-a270-d8da3bf5252a

+++ sfworks/common/PackedPtrSet.h
@@ -76,7 +76,7 @@ public:
         for(; *vp; ++vp) {
             if (*vp == p)
                 return false;
-            if (((int)(*vp)) & 01) {
+            if (((intptr_t)(*vp)) & 01) {
                 *vp = p;
                 return true;
             }
